requirements.txt
# Web Framework and CORS
Flask>=3.0.0
flask-cors==3.0.10

# Machine Learning & Data Manipulation
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.23.0
scipy>=1.10.0

# Data Validation
pydantic>=1.10.0

# Production Server
gunicorn>=21.2.0

# Google Cloud Storage Client
google-cloud-storage

# Environment Variable Loader
python-dotenv  # <-- TAMBAHKAN BARIS INI